Moving a lot of Studio analysis into Libs/Analyze
