+++ Libs/Analyze/CMakeLists.txt
@@ -1,9 +1,31 @@
+SET(ANALYZE_MOC_HDRS
+  MeshManager.h
+  QMeshWarper.h
+  MeshWorker.h
+)
+
+qt5_wrap_cpp( ANALYZE_MOC_SRCS ${ANALYZE_MOC_HDRS} )
+
 
 add_library(Analyze STATIC
   Analyze.cpp
+  CustomSurfaceReconstructionFilter.cpp
+  LegacyMeshGenerator.cpp
+  MeshCache.cpp
+  MeshGenerator.cpp
+  MeshGroup.cpp
+  MeshManager.cpp
+  MeshWorkQueue.cpp
+  MeshWorker.cpp
   Particles.cpp
+  QMeshWarper.cpp
+  Shape.cpp
+  StudioEnums.cpp
   StudioMesh.cpp
+  SurfaceReconstructor.cpp
   Vis.cpp
+  vtkPolyDataToImageData.cpp
+  ${ANALYZE_MOC_SRCS}
   )
 
 target_include_directories(Analyze PUBLIC
@@ -12,7 +34,9 @@ target_include_directories(Analyze PUBLIC
 
 target_link_libraries(Analyze PUBLIC
   Mesh
-  Utils)
+  Particles
+  Utils
+  )
 
 # set
 set_target_properties(Analyze PROPERTIES PUBLIC_HEADER
