+++ Studio/src/Data/ShapeWorksWorker.cpp
@@ -1,10 +1,10 @@
-// must come first
+// must come first due to Q_EMIT vs emit
 #include <Libs/Optimize/Optimize.h>
 // must come first
 
-#include <Data/Shape.h>
+#include <Shape.h>
 #include <Data/ShapeWorksWorker.h>
-#include <Data/SurfaceReconstructor.h>
+#include <SurfaceReconstructor.h>
 #include <Groom/QGroom.h>
 #include <Libs/Optimize/OptimizeParameters.h>
 #include <Logging.h>
