+++ Studio/src/DeepSSM/DeepSSMTool.cpp
@@ -13,7 +13,7 @@
 
 // studio
 #include <Data/Session.h>
-#include <Data/Shape.h>
+#include <Shape.h>
 #include <Data/ShapeWorksWorker.h>
 #include <Logging.h>
 #include <DeepSSM/DeepSSMJob.h>
